Move non-phockito tests out of original UriProcessorTest
