+++ tests/UnitTests/POData/UriProcessor/UriProcessorNewTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace UnitTests\POData\UriProcessor;
+
+use Mockery as m;
+
+class UriProcessorNewTest extends \PHPUnit_Framework_TestCase
+{
+    public function testTest()
+    {
+        $this->markTestSkipped('Let\'s see if autoloader picks this up');
+    }
+}
